+++ tests/run_standalone_tests.sh
@@ -16,6 +16,8 @@
 # THIS FILE ASSUMES IT IS RUN INSIDE THE tests DIRECTORY
 
 set -e
+RED='\033[0;31m'
+NC='\033[0m'
 
 # Defaults
 hpus=2
@@ -71,19 +73,20 @@ declare -a results
 # Get test list and run each test individually
 tests=$(grep -oP '^tests/test_\S+' "$TEST_FILE")
 for test in $tests; do
-  result=$(python -um pytest -sv "$test" --hpus $hpus --pythonwarnings ignore --junitxml="$test"-results.xml | tail -n 1)
+  result=$(python -um pytest -sv "$test" --hpus $hpus --pythonwarnings ignore --junitxml="$test"-results.xml)
+  retval=$?
+  last_line=$(tail -n 1 <<< "$result")
+
   pattern='([0-9]+) (.*) in ([0-9.]+s)'
   status=""
-  if [[ $result =~ $pattern ]]; then
+  if [[ $last_line =~ $pattern ]]; then
       status="${BASH_REMATCH[2]}"
-  fi
-  result="$test:$status"
-  echo $result
-  if [[ $status == "failed" ]]; then
-    cat $test-results.xml
+  elif [ "$retval" != 0 ]; then
+    echo -e "${RED}$(cat $test-results.xml)${NC}"
     exit 1
   fi
-  results+=("$result")
+
+  results+=("${test}:${status}")
 done
 
 echo "===== STANDALONE TEST STATUS BEGIN ====="