+++ tests/test_pytorch/test_compile.py
@@ -31,7 +31,7 @@ elif module_available("pytorch_lightning"):
 
 from lightning_habana.pytorch.accelerator import HPUAccelerator
 from lightning_habana.pytorch.plugins import HPUPrecisionPlugin
-from lightning_habana.pytorch.strategies import HPUParallelStrategy, SingleHPUStrategy
+from lightning_habana.pytorch.strategies import HPUDDPStrategy, SingleHPUStrategy
 
 
 @pytest.fixture()
@@ -211,15 +211,14 @@ def test_all_stages_with_compile(tmpdir, hpus):
 @pytest.mark.standalone()
 @pytest.mark.skipif(HPUAccelerator.auto_device_count() <= 1, reason="Test requires multiple HPU devices")
 @pytest.mark.usefixtures("_is_compile_allowed")
-@pytest.mark.parametrize("hpus", [2])
-def test_parallel_strategy_with_compile(tmp_path, hpus):
+def test_ddp_strategy_with_compile(tmp_path, hpus):
     """Tests compiled BoringModel on HPU."""
     model = BoringModel()
     compiled_model = torch.compile(model, backend="hpu_backend")
 
     _plugins = [HPUPrecisionPlugin(precision="bf16-mixed")]
     parallel_hpus = [torch.device("hpu")] * hpus
-    _strategy = HPUParallelStrategy(
+    _strategy = HPUDDPStrategy(
         parallel_devices=parallel_hpus,
         bucket_cap_mb=100,
         gradient_as_bucket_view=True,